pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "pipeguard-cli"
dynamic = ["version"]
description = "Pre-commit security scanner for GitHub Actions workflows"
authors = [
  { name = "Alex Jung", email = "contact@alex-jung.org" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-requests>=2.31",
]

[project.scripts]
pipeguard = "pipeguard.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["pipeguard/_version.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "pipeguard/_version.py"

[tool.hatch.build.targets.wheel]
include = [
    "pipeguard/**/*.py",
    "pipeguard/**/*.json",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
